extract random mask functions
