+++ components/Block/Block.tsx
@@ -1,5 +1,4 @@
 /** @jsxImportSource @emotion/react */
-import { css } from "@emotion/react"
 import styled from "@emotion/styled"
 import React from "react"
 import * as yup from "yup"
