+++ components/Layout/Layout.tsx
@@ -1,4 +1,4 @@
-import { css, Theme } from "@emotion/react"
+import { css } from "@emotion/react"
 import styled from "@emotion/styled"
 import React from "react"
 import { Breakpoints, min, theme } from "../../styles/theme"
